+++ src/test/java/com/dbteam/repository/PurchaseRepositoryTest.java
@@ -30,9 +30,9 @@ public class PurchaseRepositoryTest {
         Person person = new Person("Pronia", "Pronia", 1L, 2L, null);
         actualRecipient = new Person("Jan", "Jan Kowalski", 1L, 2L, null);
         List<Person> recipients = List.of(person, actualRecipient);
-        Purchase purchase1 = new Purchase(1L, "Vasia", LocalDate.now().plusDays(3), "Lewik", 10L, "", recipients);
-        Purchase purchase2 = new Purchase(2L, "Vasia", LocalDate.now(), "Lidl", 20L, "", recipients);
-        Purchase purchase3 = new Purchase(3L, "Jan", LocalDate.now().minusDays(5), "Lewik", 30L, "",
+        Purchase purchase1 = new Purchase(1L, "Vasia", LocalDate.now().plusDays(3), "Lewik", 10D, "", recipients);
+        Purchase purchase2 = new Purchase(2L, "Vasia", LocalDate.now(), "Lidl", 20D, "", recipients);
+        Purchase purchase3 = new Purchase(3L, "Jan", LocalDate.now().minusDays(5), "Lewik", 30D, "",
                 Collections.singletonList(person));
         purchaseRepository.saveAll(List.of(purchase1, purchase2, purchase3));
     }
@@ -44,9 +44,9 @@ public class PurchaseRepositoryTest {
 
         //then
         assertEquals(2, actualPurchases.size());
-        Long actualAmount = actualPurchases
+        Double actualAmount = actualPurchases
                 .stream()
-                .mapToLong(Purchase::getAmount)
+                .mapToDouble(Purchase::getAmount)
                 .sum();
         assertEquals(30, actualAmount);
     }
